pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Vitalii Madiakin <madyakinvitalii@gmail.com>"]
readme = "README.md"
include = [".env"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
spectacular = "^0.4.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
social-auth-app-django = "^5.2.0"
coreapi = "^2.3.3"
django-filter = "^23.2"
marshmallow = "^3.20.1"
marshmallow-dataclass = "^8.5.14"
requests = "^2.31.0"
python-telegram-bot = "^20.4"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
